+++ lib/legion/extensions/node.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
 require 'legion/extensions/node/version'
+require 'legion/extensions/node/helpers/rabbitmq'
 
 module Legion
   module Extensions